+++ plugin/src/net/groboclown/idea/p4ic/v2/server/util/ChangelistDescriptionGenerator.java
@@ -13,12 +13,42 @@
  */
 package net.groboclown.idea.p4ic.v2.server.util;
 
+import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vcs.changes.ChangeList;
 import net.groboclown.idea.p4ic.P4Bundle;
+import net.groboclown.idea.p4ic.config.UserProjectPreferences;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class ChangelistDescriptionGenerator {
-    public static String getDescription(@NotNull ChangeList ideaChange) {
+    public static String getDescription(@Nullable Project project, @NotNull ChangeList ideaChange) {
+        // Old behavior which isn't really that good.
+        if (UserProjectPreferences.getConcatenateChangelistNameComment(project)) {
+            StringBuilder sb = new StringBuilder();
+            if (ideaChange.getName().length() > 0) {
+                sb.append(ideaChange.getName());
+            }
+            if (ideaChange.getComment() != null && ideaChange.getComment().length() > 0) {
+                String comment = ideaChange.getComment();
+                if (sb.length() > 0) {
+
+
+                    String head = sb.toString();
+                    if (head.endsWith("...")) {
+                        head = head.substring(0, head.length() - 3);
+                    }
+                    if (head.trim().equalsIgnoreCase(comment)) {
+                        // Just use the comment, not the name + comment.
+                        sb.setLength(0);
+                    } else {
+                        sb.append("\n\n");
+                    }
+                }
+                sb.append(ideaChange.getComment());
+            }
+            return sb.toString();
+        }
+
         // Bug #91: if both "name" and "comment" are specified, then the changelist will use both
         // the name and the description for the comment.  Instead, this should match the "submit"
         // functionality and only use the comment.