+++ CHANGES.md
@@ -11,6 +11,7 @@
 
 * Bug fixes.
     * Fixed a possible deadlock with server connection checking.
+    * Fixed automatic add to VCS option fails on copy (#102) 
 
 
 ## ::v0.7.13::